changed %f to %d

+++ sum.c
@@ -4,8 +4,8 @@ int main() {
     int a = 2;
     float b = 0.5;
     double c = 0.25;
-    double sum = a + b + c;
+    int sum = a + b + c;
 
-    printf("The Sum is: %f", sum);
+    printf("The Sum is: %d", sum);
     return 0;
 }
